Added markdown and ascii render modes
I implemented an abstract set of `generate_cells` methods, following the
code paths from the `generate_html` methods. These return a
two-dimensional array of strings, which could be used for other render
formats as well. (It might be valuable to refactor the html output to
use this where possible.) The 2d grid of cells intentionally limits some
of the functionality--no rowspans or colspans are supported.

I also added a few new properties which will be valuable for anyone
trying to export markdown regression tables:

- self.precision_separator = " "
- self.df_separator = " "
- self.sig_char = '*'

The separators could be changed to '\n', which is particularly nice
because tabulate converts newlines into whatever markdown format is
requested. The sig_char may need to be replaced by an escaped version of
itself in some output contexts.

Once the `generate_cells` methods were implemented, I use `tabulate`
(new dependency) to render these cells to ascii or any supported flavor
of markdown.